Community Page WIP & Navbar fix
For the time being, I think we should take down the community page from
the website until I can take a pass at the page's design (if we need an
entire page for it) and figure out where it should be nested in the
navigation. Currently, I think that the card design it is using is kind
of confusing for this particular situation.

Ana and I were thinking that the page should give the user all of the
different means of communications (like it currently is) but displayed
different and give the user some context on what mean of communication
is used for what. If you visited it right now blindly as someone who
has never gotten in contact with the Jupyter community, it is very
confusing on where to post questions, suggestions and what each website
does for us specifically.

Also, fixed the navbar items from overflowing to the next line in the
navbar. Once I have the redesign of the community page done, we are
going to restructure the nabber to have less items displayed. We are
either going to nest items within each other (dropdown style of nav
item) or just remove some items that don’t need to be in the navbar