make it clear that Jupyter is open source
If someone just learned of Jupyter and visits the front page, it's not at all clear that we're not a company selling a product. I also added a link to IPython - since ipython.org links back here, and added the same blurb that only appears on the donate page at present about how Jupyter is and always will be BSD licensed.