More appealing Community project page.
While the docs page is nice and list (some) of the community related
project is is a bit tough to navigate for some one new to Jupyter.

This is a proposal to provide a more vibrant page, easier to get
contribution to (that is to say you do not have to write html).
Allow people to have logos, links, description, as well as a Basic
Search.

I structured with the Kernels at the top, as I believe this is really
what most people will be interested in.

I also explicitely list things like services hosting Jupyter.