Quite a bit of work and design additions.
One narrow screen this will be a single column.
Sections now have headers with logo, and items that have no provided
logo will use the generic section header.

The search is a bit better (space is used as an AND between terms), each
item get also tag with the name of the section it's in so you can filter
by section as well. The css should allow to pretty easily decide what
get shown/hidden when searching. I vent with hiding the sections
descriptions and logo (as when you search you want compact results) but
keep the sections otherwise they might be hard to discover.