Cleanup the documentation page.
The Documentation page in pure html was bothering me with lots of extra
element, and not really DRY, so hard to extend.

This fixes all that (and also use orange as the main color... why blue
we never use blue ?).

I'm also tempted to remove all the /en/latest/ or /en/stable/ from the
rtd links as RTD should detect the default page; not the pointing links,
but that will be the subject of another PR.

Also I want to change the URL from /documentation.html to
/documentation/ but same; we'll do that another time.