add run instructions to JupyterLab install page
Currently the install page shows the command to install JupyterLab, but not the command to launch it. The launch command is not "jupyterlab" (the name of the package) but rather "jupyter-lab". The proposed change adds a section below installation instructions to make it easier to discover how to launch JupyterLab after installing it.

This change also refines the text to reduce the total number of lines required to convey the same information.