Add header links
There are many times I would like to send a link to a specific part of a page, but it is not easy to get the anchor link. This adds a link available on hovering over a header, as is done on many sites.

This first pass just uses a CDN. If it works well, we might make it more self-contained.