Add a globe icon for personal website badges.
This icon comes from https://github.com/twitter/twemoji/blob/ad3d3d669bb3697946577247ebb15818f09c6c91/assets/svg/1f310.svg, licensed CC-by-4.0. In the readme, it says "We would consider a mention in the HTML/JS source sufficient also." so that is what we did.

I tweaked the color of the path to a dark grey, matching the text on the page.