Update branches names.
Master is becoming more and more rare, so this update workflow to also
work once we rename the branch. It also update a link to another repo
that now uses main. (Link it not broken as GitHub redirects, but better
to update).

Once this is merge can Someone go into settings and rename the master
branch of this repo to main ?